requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
chromadb==0.4.22
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.1.0
langchain-community==0.0.10
google-generativeai==0.3.2
openai==1.6.1
tiktoken==0.5.2
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0
Werkzeug==3.0.1
gunicorn==21.2.0
numpy==1.24.3
pandas==1.5.3